refactor(veterinary): corrigido tipo de dado do atributo weight

+++ api/src/main/java/com/ong/acolhepatinhas/api/veterinary/record/DTO/BasicVeterinaryRecordResponse.java
@@ -1,5 +1,7 @@
 package com.ong.acolhepatinhas.api.veterinary.record.DTO;
 
+import java.math.BigDecimal;
+
 import com.ong.acolhepatinhas.api.veterinary.record.VeterinaryRecord;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -13,7 +15,7 @@ public record BasicVeterinaryRecordResponse(
     Integer size,
 
     @Schema(example = "1.5", description = "Valor em kg")
-    Double weight,
+    BigDecimal weight,
 
     boolean neutered
 ) {
